ESQL: Make EvalBenchmarks executable again (#102854)
